Make DefaultRegistryFunc and PodSpec extraction generic
