+++ pkg/scheduler/admission/podnodeconstraints/admission_test.go
@@ -3,8 +3,17 @@ package podnodeconstraints
 import (
 	"bytes"
 	"fmt"
-	"reflect"
 	"testing"
+
+	admission "k8s.io/kubernetes/pkg/admission"
+	kapi "k8s.io/kubernetes/pkg/api"
+	"k8s.io/kubernetes/pkg/api/unversioned"
+	"k8s.io/kubernetes/pkg/apis/batch"
+	"k8s.io/kubernetes/pkg/apis/extensions"
+	"k8s.io/kubernetes/pkg/auth/user"
+	"k8s.io/kubernetes/pkg/runtime"
+	"k8s.io/kubernetes/pkg/serviceaccount"
+	"k8s.io/kubernetes/pkg/util/sets"
 
 	_ "github.com/openshift/origin/pkg/api/install"
 	authorizationapi "github.com/openshift/origin/pkg/authorization/api"
@@ -13,17 +22,6 @@ import (
 	deployapi "github.com/openshift/origin/pkg/deploy/api"
 	"github.com/openshift/origin/pkg/scheduler/admission/podnodeconstraints/api"
 	securityapi "github.com/openshift/origin/pkg/security/api"
-
-	admission "k8s.io/kubernetes/pkg/admission"
-	kapi "k8s.io/kubernetes/pkg/api"
-	"k8s.io/kubernetes/pkg/api/unversioned"
-	"k8s.io/kubernetes/pkg/apimachinery/registered"
-	"k8s.io/kubernetes/pkg/apis/batch"
-	"k8s.io/kubernetes/pkg/apis/extensions"
-	"k8s.io/kubernetes/pkg/auth/user"
-	"k8s.io/kubernetes/pkg/runtime"
-	"k8s.io/kubernetes/pkg/serviceaccount"
-	"k8s.io/kubernetes/pkg/util/sets"
 )
 
 func TestPodNodeConstraints(t *testing.T) {
@@ -470,90 +468,4 @@ nodeSelectorLabelBlacklist:
 	if len(config.NodeSelectorLabelBlacklist) == 0 {
 		t.Fatalf("NodeSelectorLabelBlacklist didn't take specified value")
 	}
-}
-
-func TestResourcesToCheck(t *testing.T) {
-	known := knownResourceKinds()
-	detected := kindsWithPodSpecs()
-	for _, k := range detected {
-		if _, isKnown := known[k]; !isKnown {
-			t.Errorf("Unknown resource kind %s contains a PodSpec", (&k).String())
-			continue
-		}
-		delete(known, k)
-	}
-	if len(known) > 0 {
-		t.Errorf("These known kinds were not detected to have a PodSpec: %#v", known)
-	}
-}
-
-var podSpecType = reflect.TypeOf(kapi.PodSpec{})
-
-func hasPodSpec(visited map[reflect.Type]bool, t reflect.Type) bool {
-	if visited[t] {
-		return false
-	}
-	visited[t] = true
-
-	switch t.Kind() {
-	case reflect.Struct:
-		if t == podSpecType {
-			return true
-		}
-		for i := 0; i < t.NumField(); i++ {
-			if hasPodSpec(visited, t.Field(i).Type) {
-				return true
-			}
-		}
-	case reflect.Array, reflect.Slice, reflect.Chan, reflect.Map, reflect.Ptr:
-		return hasPodSpec(visited, t.Elem())
-	}
-	return false
-}
-
-func internalGroupVersions() []unversioned.GroupVersion {
-	groupVersions := registered.EnabledVersions()
-	groups := map[string]struct{}{}
-	for _, gv := range groupVersions {
-		groups[gv.Group] = struct{}{}
-	}
-	result := []unversioned.GroupVersion{}
-	for group := range groups {
-		result = append(result, unversioned.GroupVersion{Group: group, Version: runtime.APIVersionInternal})
-	}
-	return result
-}
-
-func isList(t reflect.Type) bool {
-	if t.Kind() != reflect.Struct {
-		return false
-	}
-
-	_, hasListMeta := t.FieldByName("ListMeta")
-	return hasListMeta
-}
-
-func kindsWithPodSpecs() []unversioned.GroupKind {
-	result := []unversioned.GroupKind{}
-	for _, gv := range internalGroupVersions() {
-		knownTypes := kapi.Scheme.KnownTypes(gv)
-		for kind, knownType := range knownTypes {
-			if !isList(knownType) && hasPodSpec(map[reflect.Type]bool{}, knownType) {
-				result = append(result, unversioned.GroupKind{Group: gv.Group, Kind: kind})
-			}
-		}
-	}
-
-	return result
-}
-
-func knownResourceKinds() map[unversioned.GroupKind]struct{} {
-	result := map[unversioned.GroupKind]struct{}{}
-	for _, ka := range resourcesToCheck {
-		result[ka] = struct{}{}
-	}
-	for _, ki := range resourcesToIgnore {
-		result[ki] = struct{}{}
-	}
-	return result
 }
